feat: :rocket: Cria o componente SingleVideo que renderiza os videos
